+++ src/components/Main/Metrics/Metrics.jsx
@@ -4,7 +4,7 @@ import { BeatLoader } from "react-spinners";
 import MetricsContent from "./MetricsContent/MetricsContent.jsx";
 import MetricsEmployee from "./MetricsEmployee/MetricsEmployee.jsx";
 import MetricsStepper from "./MetricsStepper/MetricsStepper.jsx";
-import $Metrics, { $Fallback } from "./Metrics.styles.jsx";
+import $Metrics, { $Fallback, $Syf } from "./Metrics.styles.jsx";
 import { MetricsContext, STEPS } from "../../../store/metrics-context.jsx";
 import { apiCallToGetCarbonThresholds } from "../../../api/Api.jsx";
 import { appgreen } from "../../../utils/colors.styles.jsx";
@@ -53,7 +53,6 @@ export default function Metrics() {
           <BeatLoader
             color={appgreen}
             loading={true}
-            cssOverride={{ margin: "20px auto 10px auto" }}
             size={35}
             speedMultiplier={0.75}
             aria-label="Loading Spinner"
@@ -68,9 +67,8 @@ export default function Metrics() {
             stepHandler,
             thresholds: resolvedThresholds,
           };
-
           return (
-            <MetricsContext.Provider value={ctxMetrics}>
+            <MetricsContext.Provider value={ctxMetrics}>               
               <$Metrics>
                 <MetricsStepper />
                 {contentDispatcher()}
@@ -84,16 +82,14 @@ export default function Metrics() {
 }
 
 async function loader() {
-  try {
-    const thresholdsObjects = await apiCallToGetCarbonThresholds();
-    const thresholdsValues = await thresholdsObjects
-      .sort((a, b) => b.threshold - a.threshold)
+    const thresholdsObjects = await apiCallToGetCarbonThresholds()
+    .then(response => {
+      const sorted = response.sort((a, b) => b.threshold - a.threshold)
       .map((object) => object.threshold);
-    return thresholdsValues;
-  } catch (error) {
-    console.log(error);
-    //TODO
-  }
+      return sorted;
+    }).catch(error => {throw error});
+
+    return thresholdsObjects;
 }
 
 export function thresholdsLoader() {